requirements.txt
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
gevent==20.6.2
greenlet==0.4.16
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
pymongo==3.10.1
python-dateutil==2.8.1
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
urllib3==1.25.9
Werkzeug==1.0.1
zope.event==4.4
zope.interface==5.1.0
